change: package 수정, yaml 파일 추가 및 설정 코드 추가
