+++ README.md
@@ -24,7 +24,7 @@
         - list 커맨드를 통해 만들어진 바우처를 조회할 수 있다.
         - 바우처 정보를 매모리에 관리한다. 어플리케이션이 종료가 되어 데이터가 모두 사라져도 괜찮습니다. (나중에 영속성을 가지도록 변경할거에요 걱정마세요!)
 - [x]  적절한 로그를 기록하고 `logback` 설정을해서 에러는 파일로 기록된다.
-- [ ]  실행가능한 `jar` 파일을 생성한다.
+- [x]  실행가능한 `jar` 파일을 생성한다.
 
 (심화) 파일을 통한 데이터관리 기능과 고객 블랙 리스트 명단 관리기능
