+++ src/main/java/org/prgms/management/configuration/YamlPropertiesFactory.java
@@ -1,0 +1,20 @@
+package org.prgms.management.configuration;
+
+import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
+import org.springframework.core.env.PropertiesPropertySource;
+import org.springframework.core.env.PropertySource;
+import org.springframework.core.io.support.EncodedResource;
+import org.springframework.core.io.support.PropertySourceFactory;
+
+import java.io.IOException;
+
+public class YamlPropertiesFactory implements PropertySourceFactory {
+    @Override
+    public PropertySource<?> createPropertySource(String name, EncodedResource encodedResource) throws IOException {
+        var yamlPropertiesFactoryBean = new YamlPropertiesFactoryBean();
+        yamlPropertiesFactoryBean.setResources(encodedResource.getResource());
+
+        var properties = yamlPropertiesFactoryBean.getObject();
+        return new PropertiesPropertySource(encodedResource.getResource().getFilename(), properties);
+    }
+}
