+++ src/main/java/org/prgms/management/repository/VoucherRepository.java
@@ -1,6 +1,6 @@
-package org.prgms.springbootbasic.management.repository;
+package org.prgms.management.repository;
 
-import org.prgms.springbootbasic.management.entity.Voucher;
+import org.prgms.management.entity.Voucher;
 
 import java.util.Map;
 import java.util.UUID;
