+++ src/main/java/org/prgms/management/service/VoucherService.java
@@ -1,0 +1,9 @@
+package org.prgms.management.service;
+
+public interface VoucherService {
+    void run(String command);
+
+    void createVoucher();
+
+    void getVoucherList();
+}
