--- src/main/resources/application.properties
@@ -1,6 +1,0 @@
-#?? ??
-log.config.level=error
-#???? ??
-log.config.path=logs
-#???? ?
-log.config.filename=log
